+++ src/partials/Field.svelte
@@ -1,12 +1,20 @@
 <script lang="ts">
-  export let label
+  export let label = null
+  export let icon = null
 </script>
 
 <div class="flex flex-col gap-1">
-  <div class="flex justify-between">
-    <label class="font-bold">{label}</label>
-    <slot name="display" />
-  </div>
+  <slot name="label">
+    <div class="flex justify-between">
+      <label class="font-bold">
+        {#if icon}
+          <i class={`fa ${icon}`} />
+        {/if}
+        {label}
+      </label>
+      <slot name="display" />
+    </div>
+  </slot>
   <slot />
   <p class="text-sm text-gray-4">
     <slot name="info" />