+++ src/partials/FieldInline.svelte
@@ -1,12 +1,21 @@
 <script lang="ts">
-  export let label
-  export let info
+  export let label = null
+  export let icon = null
 </script>
 
 <div class="flex flex-col gap-1">
-  <div class="flex gap-2">
-    <label class="font-bold">{label}</label>
-    <slot />
-  </div>
-  <p class="text-sm text-gray-4">{info}</p>
+  <slot name="label">
+    <div class="flex justify-between">
+      <label class="font-bold">
+        {#if icon}
+          <i class={`fa ${icon}`} />
+        {/if}
+        {label}
+      </label>
+      <slot />
+    </div>
+  </slot>
+  <p class="text-sm text-gray-4">
+    <slot name="info" />
+  </p>
 </div>